Add support for MCR Prefix Filter Lists and improve MVE documentation
- Added support for MCR Prefix Filter Lists:
  - Create, list, get, update, and delete prefix filter lists on an MCR.
  - Improved output formatting for prefix filter list entries.
- Refactor and enhance error handling and output formatting:
  - Updated WrapRunE function to provide detailed error messages with additional context.
  - Applied WrapRunE to all RunE commands in mve.go, mcr.go, locations.go, and completion.go.
  - Enhanced output formatting for various commands to include detailed information.
- Improved documentation for MVE commands:
  - Added detailed descriptions for the buy and get commands.
  - Included available values for the vendor field in the buy command.
- Added tests for completion command and edge cases.
- Refactored and added utility functions for better code organization and readability.
- Updated mock services and test cases to align with the new changes.
