+++ internal/cmd/mcr_mock.go
@@ -7,48 +7,39 @@ import (
 )
 
 type MockMCRService struct {
-	GetMCRErr                          error
-	GetMCRResult                       *megaport.MCR
-	ListMCRsErr                        error
-	ListMCRsResult                     []*megaport.MCR
-	BuyMCRErr                          error
-	BuyMCRResult                       *megaport.BuyMCRResponse
-	CapturedBuyMCRRequest              *megaport.BuyMCRRequest
-	ValidateMCROrderErr                error
-	ModifyMCRErr                       error
-	ModifyMCRResult                    *megaport.ModifyMCRResponse
-	CapturedModifyMCRRequest           *megaport.ModifyMCRRequest
-	DeleteMCRErr                       error
-	DeleteMCRResult                    *megaport.DeleteMCRResponse
-	CapturedDeleteMCRUID               string
-	RestoreMCRErr                      error
-	RestoreMCRResult                   *megaport.RestoreMCRResponse
-	CapturedRestoreMCRUID              string
-	ListMCRResourceTagsErr             error
-	ListMCRResourceTagsResult          map[string]string
-	CapturedListMCRResourceTagsUID     string
-	UpdateMCRResourceTagsErr           error
-	CapturedUpdateMCRResourceTagsUID   string
-	CapturedUpdateMCRResourceTagsTags  map[string]string
-	CreatePrefixFilterListErr          error
-	CreatePrefixFilterListResult       *megaport.CreateMCRPrefixFilterListResponse
-	CapturedCreatePrefixFilterListReq  *megaport.CreateMCRPrefixFilterListRequest
-	ListMCRPrefixFilterListsErr        error
-	ListMCRPrefixFilterListsResult     []*megaport.PrefixFilterList
-	CapturedListMCRPrefixFilterListsID string
-	GetMCRPrefixFilterListErr          error
-	GetMCRPrefixFilterListResult       *megaport.MCRPrefixFilterList
-	CapturedGetMCRPrefixFilterListID   string
-	CapturedGetMCRPrefixFilterListNum  int
-	ModifyMCRPrefixFilterListErr       error
-	ModifyMCRPrefixFilterListResult    *megaport.ModifyMCRPrefixFilterListResponse
-	CapturedModifyPrefixFilterListID   string
-	CapturedModifyPrefixFilterListNum  int
-	CapturedModifyPrefixFilterListReq  *megaport.MCRPrefixFilterList
-	DeleteMCRPrefixFilterListErr       error
-	DeleteMCRPrefixFilterListResult    *megaport.DeleteMCRPrefixFilterListResponse
-	CapturedDeletePrefixFilterListID   string
-	CapturedDeletePrefixFilterListNum  int
+	BuyMCRResult                             *megaport.BuyMCRResponse
+	BuyMCRErr                                error
+	CapturedBuyMCRRequest                    *megaport.BuyMCRRequest
+	ValidateMCROrderErr                      error
+	GetMCRResult                             *megaport.MCR
+	GetMCRErr                                error
+	DeleteMCRResult                          *megaport.DeleteMCRResponse
+	DeleteMCRErr                             error
+	CapturedDeleteMCRUID                     string
+	RestoreMCRResult                         *megaport.RestoreMCRResponse
+	RestoreMCRErr                            error
+	CapturedRestoreMCRUID                    string
+	CreateMCRPrefixFilterListResult          *megaport.CreateMCRPrefixFilterListResponse
+	CreateMCRPrefixFilterListErr             error
+	CapturedCreateMCRPrefixFilterListRequest *megaport.CreateMCRPrefixFilterListRequest
+	ListMCRPrefixFilterListsResult           []*megaport.PrefixFilterList
+	ListMCRPrefixFilterListsErr              error
+	GetMCRPrefixFilterListResult             *megaport.MCRPrefixFilterList
+	GetMCRPrefixFilterListErr                error
+	ModifyMCRPrefixFilterListResult          *megaport.ModifyMCRPrefixFilterListResponse
+	ModifyMCRPrefixFilterListErr             error
+	CapturedModifyMCRPrefixFilterListRequest *megaport.MCRPrefixFilterList
+	DeleteMCRPrefixFilterListResult          *megaport.DeleteMCRPrefixFilterListResponse
+	DeleteMCRPrefixFilterListErr             error
+	ModifyMCRResult                          *megaport.ModifyMCRResponse
+	ModifyMCRErr                             error
+	CapturedModifyMCRRequest                 *megaport.ModifyMCRRequest
+	ListMCRResourceTagsResult                map[string]string
+	ListMCRResourceTagsErr                   error
+	UpdateMCRResourceTagsErr                 error
+	CapturedUpdateMCRResourceTagsRequest     map[string]string
+	GetMCRPrefixFilterListsResult            []*megaport.PrefixFilterList
+	GetMCRPrefixFilterListsErr               error
 }
 
 func (m *MockMCRService) BuyMCR(ctx context.Context, req *megaport.BuyMCRRequest) (*megaport.BuyMCRResponse, error) {
@@ -56,56 +47,18 @@ func (m *MockMCRService) BuyMCR(ctx context.Context, req *megaport.BuyMCRRequest
 	if m.BuyMCRErr != nil {
 		return nil, m.BuyMCRErr
 	}
-	if m.BuyMCRResult != nil {
-		return m.BuyMCRResult, nil
-	}
-	return &megaport.BuyMCRResponse{
-		TechnicalServiceUID: "mock-mcr-uid",
-	}, nil
+	return m.BuyMCRResult, nil
 }
 
 func (m *MockMCRService) ValidateMCROrder(ctx context.Context, req *megaport.BuyMCRRequest) error {
-	if m.ValidateMCROrderErr != nil {
-		return m.ValidateMCROrderErr
-	}
-	return nil
+	return m.ValidateMCROrderErr
 }
 
-func (m *MockMCRService) GetMCR(ctx context.Context, mcrId string) (*megaport.MCR, error) {
+func (m *MockMCRService) GetMCR(ctx context.Context, mcrUID string) (*megaport.MCR, error) {
 	if m.GetMCRErr != nil {
 		return nil, m.GetMCRErr
 	}
-	if m.GetMCRResult != nil {
-		return m.GetMCRResult, nil
-	}
-	return &megaport.MCR{
-		UID:                mcrId,
-		Name:               "Mock MCR",
-		ProvisioningStatus: "LIVE",
-	}, nil
-}
-
-func (m *MockMCRService) ListMCRs(ctx context.Context) ([]*megaport.MCR, error) {
-	if m.ListMCRsErr != nil {
-		return nil, m.ListMCRsErr
-	}
-	if m.ListMCRsResult != nil {
-		return m.ListMCRsResult, nil
-	}
-	return []*megaport.MCR{}, nil
-}
-
-func (m *MockMCRService) ModifyMCR(ctx context.Context, req *megaport.ModifyMCRRequest) (*megaport.ModifyMCRResponse, error) {
-	m.CapturedModifyMCRRequest = req
-	if m.ModifyMCRErr != nil {
-		return nil, m.ModifyMCRErr
-	}
-	if m.ModifyMCRResult != nil {
-		return m.ModifyMCRResult, nil
-	}
-	return &megaport.ModifyMCRResponse{
-		IsUpdated: true,
-	}, nil
+	return m.GetMCRResult, nil
 }
 
 func (m *MockMCRService) DeleteMCR(ctx context.Context, req *megaport.DeleteMCRRequest) (*megaport.DeleteMCRResponse, error) {
@@ -113,116 +66,77 @@ func (m *MockMCRService) DeleteMCR(ctx context.Context, req *megaport.DeleteMCRR
 	if m.DeleteMCRErr != nil {
 		return nil, m.DeleteMCRErr
 	}
-	if m.DeleteMCRResult != nil {
-		return m.DeleteMCRResult, nil
-	}
-	return &megaport.DeleteMCRResponse{
-		IsDeleting: true,
-	}, nil
+	return m.DeleteMCRResult, nil
 }
 
-func (m *MockMCRService) RestoreMCR(ctx context.Context, mcrId string) (*megaport.RestoreMCRResponse, error) {
-	m.CapturedRestoreMCRUID = mcrId
+func (m *MockMCRService) RestoreMCR(ctx context.Context, mcrUID string) (*megaport.RestoreMCRResponse, error) {
+	m.CapturedRestoreMCRUID = mcrUID
 	if m.RestoreMCRErr != nil {
 		return nil, m.RestoreMCRErr
 	}
-	if m.RestoreMCRResult != nil {
-		return m.RestoreMCRResult, nil
-	}
-	return &megaport.RestoreMCRResponse{
-		IsRestored: true,
-	}, nil
-}
-
-func (m *MockMCRService) ListMCRResourceTags(ctx context.Context, mcrID string) (map[string]string, error) {
-	m.CapturedListMCRResourceTagsUID = mcrID
-	if m.ListMCRResourceTagsErr != nil {
-		return nil, m.ListMCRResourceTagsErr
-	}
-	if m.ListMCRResourceTagsResult != nil {
-		return m.ListMCRResourceTagsResult, nil
-	}
-	return map[string]string{
-		"environment": "test",
-		"owner":       "automation",
-	}, nil
-}
-
-func (m *MockMCRService) UpdateMCRResourceTags(ctx context.Context, mcrID string, tags map[string]string) error {
-	m.CapturedUpdateMCRResourceTagsUID = mcrID
-	m.CapturedUpdateMCRResourceTagsTags = tags
-	if m.UpdateMCRResourceTagsErr != nil {
-		return m.UpdateMCRResourceTagsErr
-	}
-	return nil
+	return m.RestoreMCRResult, nil
 }
 
 func (m *MockMCRService) CreatePrefixFilterList(ctx context.Context, req *megaport.CreateMCRPrefixFilterListRequest) (*megaport.CreateMCRPrefixFilterListResponse, error) {
-	m.CapturedCreatePrefixFilterListReq = req
-	if m.CreatePrefixFilterListErr != nil {
-		return nil, m.CreatePrefixFilterListErr
+	m.CapturedCreateMCRPrefixFilterListRequest = req
+	if m.CreateMCRPrefixFilterListErr != nil {
+		return nil, m.CreateMCRPrefixFilterListErr
 	}
-	if m.CreatePrefixFilterListResult != nil {
-		return m.CreatePrefixFilterListResult, nil
-	}
-	return &megaport.CreateMCRPrefixFilterListResponse{
-		IsCreated:          true,
-		PrefixFilterListID: 123,
-	}, nil
+	return m.CreateMCRPrefixFilterListResult, nil
 }
 
-func (m *MockMCRService) ListMCRPrefixFilterLists(ctx context.Context, mcrId string) ([]*megaport.PrefixFilterList, error) {
-	m.CapturedListMCRPrefixFilterListsID = mcrId
+func (m *MockMCRService) ListMCRPrefixFilterLists(ctx context.Context, mcrID string) ([]*megaport.PrefixFilterList, error) {
 	if m.ListMCRPrefixFilterListsErr != nil {
 		return nil, m.ListMCRPrefixFilterListsErr
 	}
-	if m.ListMCRPrefixFilterListsResult != nil {
-		return m.ListMCRPrefixFilterListsResult, nil
-	}
-	return []*megaport.PrefixFilterList{}, nil
+	return m.ListMCRPrefixFilterListsResult, nil
 }
 
 func (m *MockMCRService) GetMCRPrefixFilterList(ctx context.Context, mcrID string, prefixFilterListID int) (*megaport.MCRPrefixFilterList, error) {
-	m.CapturedGetMCRPrefixFilterListID = mcrID
-	m.CapturedGetMCRPrefixFilterListNum = prefixFilterListID
 	if m.GetMCRPrefixFilterListErr != nil {
 		return nil, m.GetMCRPrefixFilterListErr
 	}
-	if m.GetMCRPrefixFilterListResult != nil {
-		return m.GetMCRPrefixFilterListResult, nil
-	}
-	return &megaport.MCRPrefixFilterList{}, nil
+	return m.GetMCRPrefixFilterListResult, nil
 }
 
 func (m *MockMCRService) ModifyMCRPrefixFilterList(ctx context.Context, mcrID string, prefixFilterListID int, prefixFilterList *megaport.MCRPrefixFilterList) (*megaport.ModifyMCRPrefixFilterListResponse, error) {
-	m.CapturedModifyPrefixFilterListID = mcrID
-	m.CapturedModifyPrefixFilterListNum = prefixFilterListID
-	m.CapturedModifyPrefixFilterListReq = prefixFilterList
+	m.CapturedModifyMCRPrefixFilterListRequest = prefixFilterList
 	if m.ModifyMCRPrefixFilterListErr != nil {
 		return nil, m.ModifyMCRPrefixFilterListErr
 	}
-	if m.ModifyMCRPrefixFilterListResult != nil {
-		return m.ModifyMCRPrefixFilterListResult, nil
-	}
-	return &megaport.ModifyMCRPrefixFilterListResponse{
-		IsUpdated: true,
-	}, nil
+	return m.ModifyMCRPrefixFilterListResult, nil
 }
 
 func (m *MockMCRService) DeleteMCRPrefixFilterList(ctx context.Context, mcrID string, prefixFilterListID int) (*megaport.DeleteMCRPrefixFilterListResponse, error) {
-	m.CapturedDeletePrefixFilterListID = mcrID
-	m.CapturedDeletePrefixFilterListNum = prefixFilterListID
 	if m.DeleteMCRPrefixFilterListErr != nil {
 		return nil, m.DeleteMCRPrefixFilterListErr
 	}
-	if m.DeleteMCRPrefixFilterListResult != nil {
-		return m.DeleteMCRPrefixFilterListResult, nil
-	}
-	return &megaport.DeleteMCRPrefixFilterListResponse{
-		IsDeleted: true,
-	}, nil
+	return m.DeleteMCRPrefixFilterListResult, nil
 }
 
-func (m *MockMCRService) GetMCRPrefixFilterLists(ctx context.Context, mcrId string) ([]*megaport.PrefixFilterList, error) {
-	return m.ListMCRPrefixFilterLists(ctx, mcrId)
+func (m *MockMCRService) ModifyMCR(ctx context.Context, req *megaport.ModifyMCRRequest) (*megaport.ModifyMCRResponse, error) {
+	m.CapturedModifyMCRRequest = req
+	if m.ModifyMCRErr != nil {
+		return nil, m.ModifyMCRErr
+	}
+	return m.ModifyMCRResult, nil
+}
+
+func (m *MockMCRService) ListMCRResourceTags(ctx context.Context, mcrID string) (map[string]string, error) {
+	if m.ListMCRResourceTagsErr != nil {
+		return nil, m.ListMCRResourceTagsErr
+	}
+	return m.ListMCRResourceTagsResult, nil
+}
+
+func (m *MockMCRService) UpdateMCRResourceTags(ctx context.Context, mcrID string, tags map[string]string) error {
+	m.CapturedUpdateMCRResourceTagsRequest = tags
+	return m.UpdateMCRResourceTagsErr
+}
+
+func (m *MockMCRService) GetMCRPrefixFilterLists(ctx context.Context, mcrID string) ([]*megaport.PrefixFilterList, error) {
+	if m.GetMCRPrefixFilterListsErr != nil {
+		return nil, m.GetMCRPrefixFilterListsErr
+	}
+	return m.GetMCRPrefixFilterListsResult, nil
 }
